+++ NEWS.md
@@ -1,5 +1,9 @@
 # webR (development version)
 
+## New features
+
+* A `webR.version` property has been added, containing the current version and build information (#409).
+
 # webR 0.3.2
 
 ## New features